# swagger.yaml
openapi: '3.0.0'
info:
  title: 'Volshebny API Documentation'
  version: '1.0.0'
  description: 'Полная документация по API для бэкенд-сервиса Volshebny. Здесь вы найдете все маршруты, их параметры и примеры ответов.'
  contact:
    name: 'Volshebny Support'
    email: 'support@volshebny.com'
servers:
  - url: 'http://localhost:5000/api'
    description: 'Локальный сервер для разработки'
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: 'Аутентификация через JWT токен в cookie. Войдите в систему через `/auth/login`, чтобы получить его.'
...
security:
  - cookieAuth: []
...
paths:
...
  /publications/{publicationId}:
    get:
      tags: [Publications]
      summary: 'Получить одну публикацию по ID'
      parameters:
        - in: path
          name: publicationId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: 'Данные публикации.' }
        '404': { description: 'Публикация не найдена.' }
    put:
      tags: [Publications]
      summary: 'Обновить свою публикацию'
      parameters:
        - in: path
          name: publicationId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                content: { type: string }
      responses:
        '200': { description: 'Публикация обновлена.' }
    delete:
      tags: [Publications]
      summary: 'Удалить свою публикацию'
      parameters:
        - in: path
          name: publicationId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: 'Публикация удалена.' }
...